remove warnings

+++ src/libawkward/builder/BoolBuilder.cpp
@@ -34,7 +34,7 @@ namespace awkward {
 
     container.copy_buffer(form_key.str() + "-data",
                           buffer_.ptr().get(),
-                          buffer_.length() * sizeof(bool));
+                          buffer_.length() * (int64_t)sizeof(bool));
 
     return "{\"class\": \"NumpyArray\", \"primitive\": \"bool\", \"form_key\": \""
            + form_key.str() + "\"}";
